move service group operations to new file
